requirements.txt
beautifulsoup4==4.9.3
lxml==4.6.2
redis==3.5.3
requests==2.25.1
rpq==2.2
rq==1.7.0
six==1.15.0
urllib3==1.26.2
selenium-wire
selenium~=3.141.0
selenium~=3.141.0
selenium-requests==1.3
PyGithub==1.55